+++ apps/decodex/src/state/models/review.rs
@@ -1,184 +1,10 @@
-/// Latest runtime-owned review-policy checkpoint for one run phase.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) struct ReviewPolicyCheckpoint {
-	pub(in crate::state) project_id: String,
-	pub(in crate::state) issue_id: String,
-	pub(in crate::state) run_id: String,
-	pub(in crate::state) attempt_number: i64,
-	pub(in crate::state) phase: String,
-	pub(in crate::state) status: String,
-	pub(in crate::state) head_sha: String,
-	pub(in crate::state) nonclean_rounds: i64,
-	pub(in crate::state) details_json: String,
-	pub(in crate::state) updated_at: String,
-	pub(in crate::state) updated_at_unix: i64,
-}
-#[cfg_attr(not(test), allow(dead_code))]
-impl ReviewPolicyCheckpoint {
-	pub(crate) fn project_id(&self) -> &str {
-		&self.project_id
-	}
+mod checkpoints;
+mod handoff;
 
-	pub(crate) fn issue_id(&self) -> &str {
-		&self.issue_id
-	}
-
-	pub(crate) fn run_id(&self) -> &str {
-		&self.run_id
-	}
-
-	pub(crate) fn attempt_number(&self) -> i64 {
-		self.attempt_number
-	}
-
-	pub(crate) fn phase(&self) -> &str {
-		&self.phase
-	}
-
-	pub(crate) fn status(&self) -> &str {
-		&self.status
-	}
-
-	pub(crate) fn head_sha(&self) -> &str {
-		&self.head_sha
-	}
-
-	pub(crate) fn nonclean_rounds(&self) -> i64 {
-		self.nonclean_rounds
-	}
-
-	pub(crate) fn details_json(&self) -> &str {
-		&self.details_json
-	}
-
-	pub(crate) fn updated_at(&self) -> &str {
-		&self.updated_at
-	}
-
-	pub(crate) fn updated_at_unix(&self) -> i64 {
-		self.updated_at_unix
-	}
-}
-
-/// Latest loop-guardrail checkpoint for one issue and stop reason.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) struct LoopGuardrailCheckpoint {
-	pub(in crate::state) project_id: String,
-	pub(in crate::state) issue_id: String,
-	pub(in crate::state) reason: String,
-	pub(in crate::state) fingerprint: String,
-	pub(in crate::state) run_id: String,
-	pub(in crate::state) attempt_number: i64,
-	pub(in crate::state) consecutive_count: i64,
-	pub(in crate::state) details_json: String,
-	pub(in crate::state) updated_at: String,
-	pub(in crate::state) updated_at_unix: i64,
-}
-impl LoopGuardrailCheckpoint {
-	#[cfg(test)]
-	pub(crate) fn project_id(&self) -> &str {
-		&self.project_id
-	}
-
-	#[cfg(test)]
-	pub(crate) fn issue_id(&self) -> &str {
-		&self.issue_id
-	}
-
-	pub(crate) fn reason(&self) -> &str {
-		&self.reason
-	}
-
-	pub(crate) fn fingerprint(&self) -> &str {
-		&self.fingerprint
-	}
-
-	pub(crate) fn run_id(&self) -> &str {
-		&self.run_id
-	}
-
-	pub(crate) fn attempt_number(&self) -> i64 {
-		self.attempt_number
-	}
-
-	pub(crate) fn consecutive_count(&self) -> i64 {
-		self.consecutive_count
-	}
-
-	pub(crate) fn details_json(&self) -> &str {
-		&self.details_json
-	}
-
-	#[cfg(test)]
-	pub(crate) fn updated_at(&self) -> &str {
-		&self.updated_at
-	}
-
-	#[cfg(test)]
-	pub(crate) fn updated_at_unix(&self) -> i64 {
-		self.updated_at_unix
-	}
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub(crate) struct ReviewHandoffMarker {
-	pub(in crate::state) run_id: String,
-	pub(in crate::state) attempt_number: i64,
-	pub(in crate::state) branch_name: String,
-	pub(in crate::state) pr_url: String,
-	pub(in crate::state) target_base_ref_name: Option<String>,
-	pub(in crate::state) pr_head_ref_name: String,
-	pub(in crate::state) pr_head_oid: String,
-}
-impl ReviewHandoffMarker {
-	pub(crate) fn new(
-		run_id: impl Into<String>,
-		attempt_number: i64,
-		branch_name: impl Into<String>,
-		pr_url: impl Into<String>,
-		target_base_ref_name: impl Into<String>,
-		pr_head_ref_name: impl Into<String>,
-		pr_head_oid: impl Into<String>,
-	) -> Self {
-		Self {
-			run_id: run_id.into(),
-			attempt_number,
-			branch_name: branch_name.into(),
-			pr_url: pr_url.into(),
-			target_base_ref_name: Some(target_base_ref_name.into()),
-			pr_head_ref_name: pr_head_ref_name.into(),
-			pr_head_oid: pr_head_oid.into(),
-		}
-	}
-
-	pub(crate) fn branch_name(&self) -> &str {
-		&self.branch_name
-	}
-
-	pub(crate) fn run_id(&self) -> &str {
-		&self.run_id
-	}
-
-	pub(crate) fn attempt_number(&self) -> i64 {
-		self.attempt_number
-	}
-
-	pub(crate) fn pr_url(&self) -> &str {
-		&self.pr_url
-	}
-
-	pub(crate) fn target_base_ref_name(&self) -> Option<&str> {
-		self.target_base_ref_name.as_deref()
-	}
-
-	pub(crate) fn pr_head_ref_name(&self) -> &str {
-		&self.pr_head_ref_name
-	}
-
-	pub(crate) fn pr_head_oid(&self) -> &str {
-		&self.pr_head_oid
-	}
-}
+pub(crate) use self::{
+	checkpoints::{LoopGuardrailCheckpoint, ReviewPolicyCheckpoint},
+	handoff::ReviewHandoffMarker,
+};
 
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub(crate) struct ReviewOrchestrationMarker {
